+++ src/libponyc/ast/ast.h
@@ -325,7 +325,7 @@ const char* ast_get_print(ast_t* ast);
 const char* ast_name(ast_t* ast);
 void ast_set_name(ast_t* ast, const char* name);
 double ast_float(ast_t* ast);
-size_t ast_int(ast_t* ast);
+__int128_t ast_int(ast_t* ast);
 ast_t* ast_type(ast_t* ast);
 void ast_settype(ast_t* ast, ast_t* type);
 void ast_erase(ast_t* ast);
